Extract nvhttp pairing module
